WebSystem/backend/requirements.txt
Django==3.2.6
django-environ==0.10.0
graphene-django==2.15.0
psycopg2-binary==2.9.1
gunicorn==19.9.0
asgiref==3.3.2
dj-database-url==0.5.0
dj-static==0.0.6
django-cors-headers==3.4.0
django-templated-mail==1.1.1
djangorestframework==3.11.1
djangorestframework-simplejwt==4.6.0
djoser==2.0.3
PyJWT==2.0.0
python-decouple==3.3
pytz==2022.7.1
sqlparse==0.4.3
static3==0.7.0
Pillow==8.1.2
dj-rest-auth==2.1.3
django-allauth==0.44.0